--- src/features/admins/admins.ts
@@ -1,4 +1,0 @@
-export * from "./constants/admins.constants";
-export * from "./constants/admins.endpoints";
-export * from "./constants/admins.paths";
-export * from "./types/admins.types";
