--- src/features/admins/components/AdminForm/AdminForm.module.scss
@@ -1,14 +1,0 @@
-@import "@/styles/variables";
-@import "@/styles/colors";
-
-.root {
-  &::before {
-    color: $text-black !important;
-    font-weight: $font-weight-700;
-    position: absolute;
-    top: 2px;
-    content: "*";
-    font-family: SimSun, sans-serif;
-    line-height: 1;
-  }
-}
