--- src/features/admins/components/ChangePasswordModal/ChangePasswordModal.module.scss
@@ -1,44 +1,0 @@
-.root {
-  :global {
-    .ant-modal {
-      &-content {
-        max-width: 572px;
-        width: 572px;
-        min-height: 249px;
-        display: flex;
-        flex-direction: column;
-        padding: 0;
-      }
-      &-header {
-        border-bottom: 1px solid #f0f0f0;
-      }
-      &-title {
-        font-size: 16px;
-        padding: 16px 24px;
-      }
-      &-body {
-        display: flex;
-        flex-direction: column;
-        justify-content: center;
-        flex: 1 !important;
-        padding: 0 24px;
-      }
-      &-footer {
-        border-top: 1px solid #f0f0f0;
-        padding: 28px 24px;
-        .ant-btn {
-          display: none;
-        }
-      }
-      .ant-form-item-label > label {
-        font-size: 14px;
-        color: #000000;
-      }
-    }
-    .btn_update_password {
-      position: absolute;
-      bottom: 3%;
-      right: 3%;
-    }
-  }
-}
