--- src/features/admins/components/CropUploadedPhoto/CropUploadedPhoto.module.scss
@@ -1,31 +1,0 @@
-.containerClass {
-  height: 130px;
-  width: 130px;
-  right: auto !important;
-  bottom: auto !important;
-  left: auto !important;
-}
-
-.controls {
-  position: absolute;
-  bottom: -160px;
-  width: 128px;
-}
-
-.zoomRange {
-  -webkit-appearance: none;
-  -moz-appearance: none;
-  appearance: none;
-  height: 2px;
-  background: #d9d9d9;
-}
-
-.zoomRange::-webkit-slider-thumb {
-  -webkit-appearance: none;
-  appearance: none;
-  border: 1px solid #d9d9d9;
-  background: #ffffff;
-  border-radius: 50%;
-  width: 8px;
-  height: 8px;
-}
