--- src/features/admins/constants/admins.endpoints.ts
@@ -1,4 +1,0 @@
-export const AdminsEndpoints = {
-  ADMINS: () => `/admins`,
-  ADMINS_DELETE: (adminID: number) => `/admins/${adminID}`,
-};
