--- src/features/admins/constants/admins.paths.ts
@@ -1,7 +1,0 @@
-export enum AdminsPathsEnum {
-  NOT_FOUND = "/404",
-  TEMPLATE = "/template",
-  ADMINS = "/admins",
-  CREATE_ADMIN = "/admins/create",
-  EDIT_ADMIN = "/admins/:adminId",
-}
