--- src/features/admins/constants/pre-sign-url.endpoints.ts
@@ -1,3 +1,0 @@
-export const PreSignUrlEndpoints = {
-  CREATE_PRE_SIGN_URL: () => `/put-presigned-url`,
-};
