--- src/features/admins/types/common.types.ts
@@ -1,6 +1,0 @@
-export type CroppedAreaPixel = {
-  width: number;
-  height: number;
-  x: number;
-  y: number;
-};
