+++ src/features/notFound/constants/notFound.paths.ts
@@ -1,0 +1,3 @@
+export enum NotFoundPathsEnum {
+  NOT_FOUND = "/404",
+}
